requirements_aethero.txt
# AETH-DEPS-001 :: AetheroOS Integration Dependencies
# Constitutional Compliance: AETH-CONST-2025-001

# Core FastAPI and async dependencies (from Superagent)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
asyncio-mqtt>=0.14.0

# Database and storage
prisma>=0.11.0
psycopg2-binary>=2.9.7
chromadb>=0.4.15
weaviate-client>=3.25.0

# Vector embeddings and ML
sentence-transformers>=2.2.2
numpy>=1.24.3
scikit-learn>=1.3.0

# Interface and visualization 
gradio>=4.0.0
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0

# WebSocket and real-time communication
websockets>=11.0.2
python-socketio>=5.9.0

# Logging and monitoring
colorlog>=6.7.0
structlog>=23.2.0
prometheus-client>=0.19.0

# Configuration and environment
python-decouple>=3.8
pyyaml>=6.0.1
toml>=0.10.2

# Data processing and validation
pandas>=2.1.0
jsonschema>=4.19.0
marshmallow>=3.20.0

# Cryptography for constitutional compliance
cryptography>=41.0.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
isort>=5.12.0
mypy>=1.6.0

# Optional dependencies for enhanced functionality
# Uncomment as needed:

# Advanced vector databases
# pinecone-client>=2.2.4
# qdrant-client>=1.6.0

# Additional ML/AI capabilities
# openai>=1.0.0
# anthropic>=0.7.0
# together>=0.2.0

# Advanced monitoring
# grafana-client>=3.5.0
# datadog>=0.47.0

# Production deployment
# gunicorn>=21.2.0
# docker>=6.1.0
# kubernetes>=27.2.0
